requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aggdraw==1.3.16
aiohttp @ file:///opt/concourse/worker/volumes/live/b3617caf-3b05-41e0-5120-f62332fd1317/volume/aiohttp_1646806401506/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.7
astunparse==1.6.3
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
cffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_62rp5d8fd4/croots/recipe/cffi_1659598655556/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///opt/concourse/worker/volumes/live/2d66025a-4d79-47c4-43be-6220928b6c82/volume/click_1646056610594/work
cmake==3.25.0
colourmap==1.1.7
cryptography @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b470f7cb-c8f1-42c9-b84f-23789ea77e7c9difcqh4/croots/recipe/cryptography_1652101134392/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==0.29.32
dcor==0.5.6
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
entrypoints==0.4
etils==0.7.1
executing==0.9.1
fastjsonschema==2.16.1
fdasrsf==2.3.11
filelock==3.8.0
findiff==0.9.2
flatbuffers @ file:///home/ktietz/cip/python-flatbuffers_1634039120618/work
fonttools==4.25.0
frozenlist @ file:///Users/builder/ci_310/frozenlist_1642502209994/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
gdown==4.5.1
giotto-ph==0.2.2
giotto-tda==0.6.0
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib==0.4.1
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
googleapis-common-protos==1.56.4
GPy==1.10.0
graphviz==0.20.1
grpcio @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0b5bbb80-272e-49e9-bc1f-f481ea3f8a681bk1a1l5/croots/recipe/grpcio_1655801042262/work
h5py @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_7fnj4n39p5/croots/recipe/h5py_1659091379933/work
hdbscan==0.8.28
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
igraph==0.10.1
imageio==2.21.0
importlib-metadata @ file:///opt/concourse/worker/volumes/live/a8740f82-0523-4b08-5bb5-afa0c929f5e0/volume/importlib-metadata_1648562424930/work
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1637872120405/work/keras-2.7.0-py2.py3-none-any.whl
Keras-Applications==1.0.8
keras-facenet==0.3.2
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
keras-vggface @ git+https://github.com/yaledhlab/vggface.git@b76539b7588bca69b0030ad7e4f985f877dc7c0a
kiwisolver @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ca2945cc-8f2f-407b-8e24-f05d29ee2f4fvto5sdlk/croots/recipe/kiwisolver_1653292053344/work
kmapper==2.0.1
llvmlite==0.39.0
lmmaes==0.3.2
lxml==4.9.1
Markdown @ file:///Users/builder/ci_310/markdown_1642534091530/work
MarkupSafe==2.1.1
matplotlib @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_croot-hkd3cnd2/matplotlib-suite_1647506470899/work
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///opt/concourse/worker/volumes/live/a9eca864-069c-4240-73c8-4e51b036a10d/volume/mkl_random_1639994047876/work
mkl-service==2.4.0
mpmath==1.2.1
mtcnn==0.1.1
multidict @ file:///Users/builder/ci_310/multidict_1642534277559/work
multimethod==1.8
munkres==1.1.4
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numba==0.56.0
numpy==1.22.4
oauthlib==3.1.0
opencv-python==4.6.0.66
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.3
pandocfilters==1.5.0
paramz==0.9.5
parso==0.8.3
patsy==0.5.2
pca==1.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.10.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==1.4.2
pydotplus==2.0.2
pyflagser==0.4.5
Pygments==2.12.0
PyJWT @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_eec47dd9-fcc8-4e06-bbca-6138d11dbbcch6zasfc4/croots/recipe/pyjwt_1657544589510/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent==0.18.1
PySocks @ file:///Users/builder/ci_310/pysocks_1642536366386/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.2.0
rdata==0.8
requests @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_516b78ce-034d-4395-b9b5-1d78c2847384qtnol99l/croots/recipe/requests_1657734628886/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scatterd==1.1.3
scikit-datasets==0.2.1
scikit-fda==0.7.1
scikit-learn==1.1.2
scipy @ file:///Users/builder/ci_310/scipy_1642604663510/work
seaborn==0.12.0
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
sympy==1.10.1
tenacity==8.0.1
tensorboard @ file:///tmp/build/80754af9/tensorboard_1633093581375/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///Users/builder/ci_310/tensorboard-data-server_1642537836006/work/tensorboard_data_server-0.6.0-py3-none-macosx_10_9_x86_64.whl
tensorboard-plugin-wit==1.6.0
tensorflow @ file:///Users/uwe/Development/conda-forge/tensorflow-feedstock/tensorflow-feedstock-osx_64_numpy1.21python3.10.____cpython/miniforge3/conda-bld/tensorflow-split_1644027665818/work/tensorflow_pkg/tensorflow-2.7.0-cp310-cp310-macosx_10_9_x86_64.whl
tensorflow-addons==0.18.0
tensorflow-datasets==4.6.0
tensorflow-estimator @ file:///Users/uwe/Development/conda-forge/tensorflow-feedstock/tensorflow-feedstock-osx_64_numpy1.21python3.10.____cpython/miniforge3/conda-bld/tensorflow-split_1644027665818/work/tensorflow-estimator/wheel_dir/tensorflow_estimator-2.7.0-py2.py3-none-any.whl
tensorflow-metadata==1.9.0
termcolor==1.1.0
terminado==0.15.0
texttable==1.6.4
tf-slim==1.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tornado @ file:///Users/builder/ci_310/tornado_1642538510757/work
tqdm==4.64.0
traitlets==5.3.0
typeguard==2.13.3
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_53dwuk2xkk/croots/recipe/urllib3_1659110467968/work
visualkeras==0.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
wget==3.2
widgetsnbextension==3.6.1
wrapt @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_1ade1f68-8354-4db8-830b-ff3072015779vd_2hm7k/croots/recipe/wrapt_1657814407132/work
xarray==2022.6.0
yarl @ file:///Users/builder/ci_310/yarl_1642513699997/work
zipp @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b279673d-f037-44c7-8773-c5a6b6f51037d3wfr9cq/croots/recipe/zipp_1652341773612/work
